Fixes #182
