+++ src/components/Modals/ValidationReportModal.vue
@@ -78,7 +78,8 @@
 <script>
 import { SHACL_URI } from "../../util/constants";
 import { urlToName } from "../../util/urlParser";
-import { capitalizeFirstLetter, groupBy } from "../../util";
+import { groupBy } from "../../util";
+import {capitalizeFirstLetter} from "../../util/strings";
 
 export default {
   name: "ValidationReportModal",
